+++ src/components/comments-markup.js
@@ -1,0 +1,23 @@
+import {formatDate} from '../utils.js';
+
+const createCommentsMarkup = (comments) => {
+  return comments.map((comment) => {
+    return (
+      `<li class="film-details__comment">
+        <span class="film-details__comment-emoji">
+          <img src="./images/emoji/${comment.emotion}" width="55" height="55" alt="emoji-smile">
+        </span>
+        <div>
+          <p class="film-details__comment-text">${comment.text}</p>
+          <p class="film-details__comment-info">
+            <span class="film-details__comment-author">${comment.author}</span>
+            <span class="film-details__comment-day">${formatDate(comment.date)}</span>
+            <button class="film-details__comment-delete">Delete</button>
+          </p>
+        </div>
+      </li>`
+    );
+  }).join(`\n`);
+};
+
+export default createCommentsMarkup;
