+++ src/components/film-popup.js
@@ -1,26 +1,7 @@
-import {formatDuration, formatDate} from '../utils.js';
+import {formatDuration} from '../utils.js';
 import createGenresMarkup from './genres-markup.js';
 import createDetailRowsMarkup from './detail-rows-markup.js';
-
-const createCommentsMarkup = (comments) => {
-  return comments.map((comment) => {
-    return (
-      `<li class="film-details__comment">
-        <span class="film-details__comment-emoji">
-          <img src="./images/emoji/${comment.emotion}" width="55" height="55" alt="emoji-smile">
-        </span>
-        <div>
-          <p class="film-details__comment-text">${comment.text}</p>
-          <p class="film-details__comment-info">
-            <span class="film-details__comment-author">${comment.author}</span>
-            <span class="film-details__comment-day">${formatDate(comment.date)}</span>
-            <button class="film-details__comment-delete">Delete</button>
-          </p>
-        </div>
-      </li>`
-    );
-  }).join(`\n`);
-}
+import createCommentsMarkup from './comments-markup.js';
 
 const createFilmPopup = (film) => {
   const {name, src, rating, release, duration, genres, description, comments, age, director, writers, actors, country} = film;
